requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-multipart
passlib
bcrypt
pytest
httpx
python-jose
google-api-python-client
google-cloud-translate